Anom Pages Setup
The following pages were created in the anom folder alongside with their components:
- Home, 
- Shop,
- Contacts,
- Basket

Some of the main layout components were created such as Navigation (+ Sidebar), Header and Footer. 

As for the shared components we have new files (+modified ones) :
- Button,  RoundImage, Text, Underline

At this stage we can say that we have the foundations to continue further and develop the backend for the application. 
-